Correctly cast the extension lengths - backport of (#2144)
    Correctly use a 32 bit integer for accumulating the length of ping
    extensions.

    The current code may accidentally truncate the length of an
    extension that is greater than 64kb and fail the validation check. We
    don't currently emit any extensions that are this length, but if we were
    to do so in the future we might have issues with older nodes (without
    this fix) will silently drop packets from newer nodes. We should
    backport this to all versions.

+++ src/cluster.c
@@ -2801,7 +2801,7 @@ int clusterProcessPacket(clusterLink *link) {
         if (hdr->mflags[0] & CLUSTERMSG_FLAG0_EXT_DATA) {
             clusterMsgPingExt *ext = getInitialPingExt(hdr, count);
             while (extensions--) {
-                uint16_t extlen = getPingExtLength(ext);
+                uint32_t extlen = getPingExtLength(ext);
                 if (extlen % 8 != 0) {
                     serverLog(LL_WARNING, "Received a %s packet without proper padding (%d bytes)",
                         clusterGetMessageTypeString(type), (int) extlen);
